Changed to getting sources from request body to appsettings.json, because it makes more sense that way.
Changed link fields in models from Uri properties to string to allow for them to be optionally set.

Changed NewsRssParser.GetNewsItems Linq query to allow for non-required fields to be optionally set depending on if the rss feed suppies them.

Changed NewsRssParser.SetNewsSourceIamgeUrl to not set the ImageUrl if one is not provided by the rss feed.

Changed ordering of news items to descending by default, as it makes more sense to have the latest news at the top. The order can now be set to ascending by using the ?dateSort=ascending query parameter.

Set null fields to be removed from json response.

Started adding logging and exception handling.